aurelium-nexus/apps/api/requirements.txt
fastapi==0.100.0
uvicorn==0.23.2
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.0
httpx==0.24.1
web3==6.4.0
pyjwt==2.8.0
eth-account==0.8.1
